servo: Merge #8987 - Report CSS errors to script task for further processing (from jdm:devtoolsreport); r=glennw
This was a missing piece from #8838.

Source-Repo: https://github.com/servo/servo
Source-Revision: b870aa90878acf97074fa55260beb8acfd125f22

UltraBlame original commit: 850c13b1564b73c3729de27a994a9a3b46faec00

+++ servo/components/script_traits/lib.rs
@@ -147,7 +147,7 @@ pub enum ConstellationControlMsg {
     
     FramedContentChanged(PipelineId, SubpageId),
     
-    ReportCSSError(PipelineId, String, u32, u32, String),
+    ReportCSSError(PipelineId, String, usize, usize, String),
 }
